+++ lib/di/service_module.dart
@@ -1,5 +1,6 @@
 import 'dart:developer';
 
+import 'package:alice/alice.dart';
 import 'package:dio/dio.dart';
 import 'package:injectable/injectable.dart';
 
@@ -7,13 +8,10 @@ import '../service/pixabay_service.dart';
 
 @module
 abstract class ServiceModule {
-  Dio getDio() {
-    var dio = Dio(BaseOptions(contentType: "application/json"));
-    _initializeInterceptors(dio);
-    return dio;
-  }
 
-  _initializeInterceptors(Dio dio) {
+  @dev
+  Dio getDevDio(Alice alice) {
+    var dio = Dio(BaseOptions(contentType: "application/json"));
     dio.interceptors.add(
       InterceptorsWrapper(
         onRequest: (options, handler) {
@@ -22,7 +20,25 @@ abstract class ServiceModule {
         },
       ),
     );
+    dio.interceptors.add(alice.getDioInterceptor());
+    return dio;
   }
+
+  @prod
+  Dio getProDio() {
+    var dio = Dio(BaseOptions(contentType: "application/json"));
+    dio.interceptors.add(
+      InterceptorsWrapper(
+        onRequest: (options, handler) {
+          options.queryParameters.putIfAbsent("key", () => PixabayService.KEY);
+          return handler.next(options);
+        },
+      ),
+    );
+
+    return dio;
+  }
+
 
   PixabayService getService(Dio dio) => PixabayService(dio);
 }